% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_AA.R
\name{codon_to_AA}
\alias{codon_to_AA}
\title{Translation to Aminoacid sequence}
\usage{
codon_to_AA(codons)
}
\arguments{
\item{codons}{}
}
\value{
AA_seq
}
\description{
Translation to Aminoacid sequence
}
\examples{
codon_to_AA("UAG")
codon_to_AA(c("UUU","GAA"))
}
