% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codons.R
\name{creates_codons}
\alias{creates_codons}
\title{Identify the codons in the RNA sequence}
\usage{
creates_codons(RNA, start = 1)
}
\arguments{
\item{RNA}{insert the RNA sequence}

\item{start}{specify the start of the sequence}
}
\value{
codons
}
\description{
Identify the codons in the RNA sequence
}
\examples{

creates_codons("AUGCUUACGUCA")

}
