% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DNA.R
\name{DNA_seq}
\alias{DNA_seq}
\title{Generate DNA sequence}
\usage{
DNA_seq(seq_size)
}
\arguments{
\item{seq_size}{}
}
\value{
complete_seq
}
\description{
Generate DNA sequence
}
\examples{

generate_DNA(125)
}
