% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DNA.R
\name{generate_DNA}
\alias{generate_DNA}
\title{Random DNA sequence generator}
\usage{
generate_DNA(seq_size)
}
\arguments{
\item{seq_size}{}
}
\value{
complete_seq
}
\description{
Random DNA sequence generator
}
\examples{

generate_DNA(125)
}
