% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_from_seq.R
\name{hist_from_seq}
\alias{hist_from_seq}
\title{The function sums counts of unique symbols found in the provided sequence and returns histogram.}
\usage{
hist_from_seq(seq)
}
\arguments{
\item{seq}{}
}
\value{
histogram
}
\description{
The function sums counts of unique symbols found in the provided sequence and returns histogram.
}
\examples{

hist_from_seq('ATTTTAACCGGCCCAA')


}
