% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_to_U.R
\name{T_to_U}
\alias{T_to_U}
\title{Translation of DNA into RNA}
\usage{
T_to_U(DNA)
}
\arguments{
\item{DNA}{}
}
\value{
RNA
}
\description{
Translation of DNA into RNA
}
\examples{

T_to_U("ACTUATAUATAUATA")

}
