% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription of DNA into RNA}
\usage{
transcription(DNA)
}
\arguments{
\item{DNA}{}
}
\value{
RNA
}
\description{
Transcription of DNA into RNA
}
\examples{

transcription("ACTUATAUATAUATA")

}
