% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation to Aminoacid sequence}
\usage{
translation(codons)
}
\arguments{
\item{codons}{}
}
\value{
AA_seq
}
\description{
Translation to Aminoacid sequence
}
\examples{
translation("UAG")
translation(c("UUU","GAA"))
}
